+++ MachOKit/Segments/MKSegment.m
@@ -282,7 +282,7 @@
         case MKNodeVMAddress:
             return _vmAddress;
         default: {
-            NSString *reason = [NSString stringWithFormat:@"Invalid node address type [%" PRIuPTR "].", type];
+            NSString *reason = [NSString stringWithFormat:@"Invalid node address type [%lu].", (unsigned long)type];
             @throw [NSException exceptionWithName:NSInvalidArgumentException reason:reason userInfo:nil];
         }
     }
